Refactor cloudflare-rs cli to be called from http.rs (#841)
* Refactor cloudflare-rs cli to call from http.rs

* pass functions into format_error for help annotations

* s/todo/TODO

* get rid of unused enum

* get rid of empty string logic in favor of Some/None
